class seats updated

+++ src/pages/Dasboard/Student/MySelectedClasses.jsx
@@ -1,17 +1,30 @@
 import { Link } from "react-router-dom";
 import useSelectedClass from "../../../hooks/useSelectedClass";
+import Swal from "sweetalert2";
 
 const MySelectedClasses = () => {
   const [selectedClass, refetch] = useSelectedClass();
   const handleDelete = (id) => {
-    fetch(`http://localhost:5000/selectedClass/${id}`, { method: "DELETE" })
-      .then((res) => res.json())
-      .then((data) => {
-        refetch();
-        console.log(data);
-      });
-
-    console.log(id);
+    Swal.fire({
+      title: "Are you sure?",
+      text: "You won't be able to revert this!",
+      icon: "warning",
+      showCancelButton: true,
+      confirmButtonColor: "#3085d6",
+      cancelButtonColor: "#d33",
+      confirmButtonText: "Yes, delete it!",
+    }).then((result) => {
+      if (result.isConfirmed) {
+        fetch(`http://localhost:5000/selectedClass/${id}`, { method: "DELETE" })
+          .then((res) => res.json())
+          .then((data) => {
+            refetch();
+            if (data.deletedCount > 0) {
+              Swal.fire("Deleted!", "Your file has been deleted.", "success");
+            }
+          });
+      }
+    });
   };
   return (
     <div className="overflow-x-auto">
